[INT-204] Replace URL search calendar state with history state + TanStack Query
